requirements.txt
Django==1.5.1
South==0.7.6
argparse==1.2.1
django-mailgun==0.2.1
gunicorn==0.17.2
httplib2==0.8
iso8601==0.1.4
oauth2==1.5.211
python-twitter==0.8.7
requests==1.2.0
simplejson==3.1.3
six==1.3.0
twilio==3.4.5
unittest2==0.5.1
wsgiref==0.1.2
